Migration to spring-boot
