+++ web/src/main/java/com/epam/esm/GiftCertificateSpringBootRestApplication.java
@@ -1,0 +1,27 @@
+package com.epam.esm;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.support.ResourceBundleMessageSource;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@SpringBootApplication
+public class GiftCertificateSpringBootRestApplication implements WebMvcConfigurer {
+    
+    private final static String LOCALIZATION_PATCH = "localization/error_messages";
+
+    public static void main(String[] args) {
+        SpringApplication.run(GiftCertificateSpringBootRestApplication.class);
+    }
+
+    @Bean
+    public ResourceBundleMessageSource messageSource() {
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setBasename(LOCALIZATION_PATCH);
+        messageSource.setDefaultEncoding("UTF-8");
+        messageSource.setUseCodeAsDefaultMessage(true);
+        return messageSource;
+    }
+
+}
