+++ web/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/epam_lab_task_2?useSSL=false
+spring.datasource.username=epam
+spring.datasource.password=epam
