--- web/src/main/resources/mysql.properties
@@ -1,4 +1,0 @@
-mysql.url=jdbc:mysql://localhost:3306/epam_lab_task_2?useSSL=false
-mysql.driver=com.mysql.cj.jdbc.Driver
-mysql.user=epam
-mysql.password=epam
